tools: new tool to get gcov data out of dirty firmware
With the new feature to download the used firmware out of sysfs and
a way to get GCOV data into that. It is now time to extract this
information and generate .gcda files for higher level tools to work
with.
Instead of dealing with the details of that just use what the
compiler gives us when we link to its gcov lib.

Signed-off-by: Henning Schild <henning.schild@siemens.com>
[Jan: clean .gcda files, fix usage of error(),
      fix signed/unsigned comparison]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>