core, inmates: Export CONFIG vars to simplify config.mk inclusion
By exporting every CONFIG variable we add to config.mk, we can pull the
inclusion of that file to higher levels, reducing the need to add it on
a per-arch basis. We still cannot include it at the top level because
the driver has to use the kernel's config, not our own.

This makes config.mk hypervisor and inmate-wide available.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>