core: Rework architecture subdir build
Avoid the recursive make invocation just to express the dependency of
the arch tree on the generated asm-defines.h. We can do better than
that.

The trick is to let the subdir depend on the asm-defines.h, and the
built-in.o on the subdir, with proper obj prefixes. But just like for
the generated files, do not create the dependency if we are cleaning up,
that would fail.

Rename the defines-file variable at this chance to something more
meaningful.

We can furthermore clean up the dependencies: asm-defines.h is for
assembly only, and that lives in the arch subdir. So there is no need to
make the CORE_OBJECTS depend on that header.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>