driver: (request|release)_mem_region for hypervisor memory
Using these kernel functions we can officially mark the hypervisor memory as
such. It will create a new entry in /proc/iomem and it will fail if another
kernel subsystem already claimed that memory. i.e. if the user did not specify
memmap on x86 or did not modify the device tree on ARM.
We could do the same for inmate memory, at least the parts we load. But that
would be a little more complex. For now that will cover the common case where
the reservation is simply missing.

Signed-off-by: Henning Schild <henning.schild@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>