x86: Account for read-only GDT in Linux 4.12
So far we were adjusting the TSS descriptor of Linux in-place during
shutdown so that we can reload TR from it. However, kernel 4.12 started
to write-protect the GDT. This causes a page fault in arch_cpu_restore.

We could now either push TR reloading into the driver module, adding
more arch-specifics to it, or use our own GDT for the TR loading step.
This patch chooses the latter.

This implies expanding our GDT to Linux size, still checking during init
for any potential overflow.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>