Provide the inmate library under BSD-2 license too.
This patch does not introduce any functional change, but "only" deals
with copyright and license aspects.

The rationale is simply that, as was already stated in COPYING, cell
inmates, AKA guests, are not derived work of Jailhouse, even if written
exclusively to run in its special environment. The inmate library
provides a set of functionalities to let inmates run on Jailhouse.
In order to avoid that people need to rewrite such library, possibly
introducing bugs or incompatibilities this way, we should rather enable
the reuse of the "official" one. Thus, it should be provided also under a
permissive license. BSD 2-clause is a well-established and easy-to-deal
with license, already used for the Jailhouse headers.

All required contributors to the affected code have signed-off to
confirm that they agree with this relicensing. Commits 98000ec9e5da
("inmate: x86: fixup apic_freq to be unsigned long") and 314922e8bb72
("inmates/lib: cmdline.c") are trival changes/removals and do not block
the relicensing.

Signed-off-by: Claudio Scordino <claudio@evidence.eu.com>
Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>
Signed-off-by: Valentine Sinitsyn <valentine.sinitsyn@gmail.com>
Signed-off-by: ShengYan Penguin <peter03192003@gmail.com>
Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Signed-off-by: Andreas Kölbl <andreas.koelbl@st.oth-regensburg.de>
Signed-off-by: Vitaly Andrianov <vitalya@ti.com>
Signed-off-by: Antonios Motakis <antonios.motakis@huawei.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com> (for all Siemens commits)