driver: handle a NULL pointer reference on the error path
In the function jailhouse_hypervisor_enable(), jumping to the label
error_unmap results in a call to jailhouse_free_firmware() which
releases hypervisor_mem_res and sets it to NULL. However, the
execution proceeds to the label error_release_memreg and tries to
access hypervisor_mem_res->start, which triggers a NULL pointer
reference.

Fix the problem for explicitly checking against the NULL pointer.

Signed-off-by: Göktürk Yüksek <gokturk@binghamton.edu>
[Jan: tune wording of comment]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>