docs, configs: Make passive comm region read-only
Specify in the docs and implement in the configuration files that cells
with passive comm regions have only read access to it. This avoids that
a passive cell, i.e. one that should also not be able to prevent
configuration changes, can block them by declaring itself "running/
locked".

Instead of checking in the hypervisor that a cell config does not permit
writes to passive comm regions, we should eventually validate this
offline via a config checker.

Reported-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>