arm: Declare each redistributor region to be last
As long as we are virtualizing MPIDR and, thus, also the region layout,
we can easily mark the end. However, when removing this virtualization
and starting to hand out the regions as-is, it will become complex to
identify contiguous ones.

To make this simple (for the hypervisor), just declare each of them to
be last. That means we will have to construct device trees for non-root
cells that declare multiple redistributor regions, even if they are
physically contiguous.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>