arm-common: Refactor handle_sgir_access to handle_dist_access
This prepares for having both GICv2 and v3 code compiled in:

Move more version-specific distributor access handling into a callback.
Specifically, we need to account for the different addresses of the
ID registers. This involves a little bit of code duplication, but
handling the difference in gic_handle_dist_access would not be simpler.

Also rename the ID register constants to make the difference clearer.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>