arm-common: Select GIC version during runtime
Unconditionally compile in GIC v2 and, on ARM, v3. The irqchip
implementation is selected in irqchip_init based on the system
configuration, and the corresponding struct irqchip is copied into a
local copy that provides quick access during runtime.

This requires to add a runtime check to gicv3_handle_sgir_write in case
it is called in v3 mode and report this back to the caller.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>