arm-common: Add GIC redistributor access filtering
Do not allow the cells to access all registers of the redistributor,
specifically not unknown ones for writing.

Also prevent accesses to redistributors that belong to other cells. The
root cell was able to do this so far as 5db9788ff2c8 didn't unmap
regions that were actually assigned to non-root cells. Soon, all cells
will see all regions, but this filter will only permit safe reads to
selected registers.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>