Stop waiting for message reply after a certain amount of time
A buggy/misbehaving inmate can hang the root cell by never replying to a
request. That's not good for fault tolerance. We therefore wait a
certain amount of time after which we consider the cell as failing.

Since the hypervisor does not currently have any notion of time (to
check the amount of elapsed time waiting), we just wait a certain amount
of loops.  The number of loops is set through the "msg_reply_timeout"
field in the cell config. By default, the timeout mechanism is disabled.
We will likely want to go from loops to a proper timing unit if the
hypervisor will get support for timers.

Signed-off-by: Claudio Scordino <claudio@evidence.eu.com>
[Gustavo Lima Chaves: Fix header format in jailhouse-cell-linux]
[Jan: minor wording and style adjustment]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>