x86,mmio: Add support for 1 more instructions on the MMIO dispatcher
Now comes

X86_OP_MOV_IMMEDIATE_TO_MEM                     0xc7,

a instruction with an immediate value encoded in. The reason to bring it
in is that the code is simple enough and will enable one more guest
OS (Zephyr OS) to function in xAPIC mode as well.

Because we now have to keep state from a contiguous memory region (the
immediate value) and an instruction may fall in between two pages in the
guest, there is extra code handling that.

Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>
Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>
[Jan: declare loop counter earlier and as unsigned]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>