x86: Relocate parking code page
This step prepares for configurable cell reset addresses. In order to
enable them, we need a stable parking code address so that we do not
need to generate per-cell mappings for this purpose.

Simply place the parking loop at address 0. Use a regular SIPI vector 0
for resetting the affected CPUs. This means on Intel that we have to
issue the DEBUGCTL reset separately in vcpu_park because we no longer
take the APIC_BSP_PSEUDO_SIPI path.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>