x86: Export PCI mmconfig base address to Linux inmates
In order to access the PCI express extended configuration space, a linux
inmate needs to have access to the PCI mmconfig base address. As inmates
doesn't have access to this information using ACPI, we need to export it
using the cell comm_region.

Note that the updates in Linux kernel used as inmate are also necessary
in order to use the exported value as the PCI mmconfig base address.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>