x86: Fix general protection fault on restoring TR
arch_cpu_restore can run concurrently on multiple CPUs, but our
modifications to the GDT in order to reload Linux TR requires that a
couple of instructions run atomically. Namely, loading TR will set the
busy flag in the descriptor again which may race with another CPU just
trying to load from that descriptor as well. Putting everything under a
spinlock resolves the race.

The effect of this bug were sporadic general protection faults (#13)
while disabling Jailhouse or when enabling it failed.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>