x86: Ignore accesses to unassigned IOAPIC pins
Do not stop a cell CPU that accesses an IOAPIC redirection table entry
for a pin that it does not own. Rather ignore the access, returning 0 on
reads. This behavior is in line with how we handle GIC accesses on ARM.
It allows guests like Linux to keep their usual IOAPIC initialization
code as-is which may first of all masks all pins of the chip.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>