inmates, tools: Avoid warnings of objtool
Latest 4.15 kernel comes with the ORC unwinder and, along that, a
number of build-time checks on object files. These are tailored for the
kernel and do not go well with both our inmates and the userspace tools,
e.g.:

.../jailhouse.o: warning: objtool: read_file() falls through to next function parse_cell_id()
.../jailhouse.o: warning: objtool: help() falls through to next function open_dev()
.../jailhouse.o: warning: objtool: open_dev() falls through to next function cell_simple_cmd()
(because of lacking knowledge about no-return functions)

Disable them for x86 inmates, currently the only affected arch, and the
tools.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>