inmates: x86: Enhance Linux setup data structure
Expand the Jailhouse-specific setup data structure we pass to the Linux
kernel on boot by the calibrated TSC and APIC timer frequencies. Also
switch the type to a value we are going to reserve officially in the
upstream kernel, rather than using the large magic value.

Although this is a change that should normally increase the version of
the struct, we keep it untouched as we didn't release the previous
structure version yet. Specifically, there is no official kernel out
there expecting the old structure.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>