configs: Fix AMD Seattle configurations
The Seattle configs had a serious bug: they granted the root cell full
access to the hypervisor memory. This caused spurious lock-ups and broke
the virtual hypervisor console consistently.

Fix this and also reorganize the layout. For this, one has to know that
the root cell Linux apparently has some business with the top most RAM.
So, leave the last 256M to that cell and put the hypervisor and IVSHMEM
below that. Lower down, we can place the non-root cell and the bare-
metal demos.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>