tools: cell-linux: Make kernel decompression factor configurable
The required factor depends on the chosen compression method, and that
may vary. Have a large factor to account for aggressive compressions
(and increased memory needs during decompression) will work - as long as
there is enough memory assigned to the cell. Using a smaller default
will address this but break if the user chooses a better compression
method.

Let's make this factor configurable in order to give the user some
control in case our default should not work.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Reviewed-by: Lokesh Vutla <lokeshvutla@ti.com>