core: Move temporary mapping area into CPU-local region
Move the area where we temporarily map guest memory in front of the
per-cpu region. This makes the mapping private for each CPU, no other
can peek into it anymore.

This change also brings in a number of simplification, starting with the
calculation of the mapping address which is now static. Furthermore, we
no longer need to reserve space in the common remapping area. And then
we can share a generic LOCAL_CPU_BASE definition via jailhouse/paging.h.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>