driver: Avoid dependency on additional kernel symbol exports
If CONFIG_KALLSYMS_ALL is enabled, we can get symbol addresses that the
upstream kernel does not export via kallsyms_lookup_name. We can use
this to avoid a number of kernel patches. Specifically, x86 can become
patch-free once all guest-side support is merged and ivshmem-net is
built as external module (or merged as well).

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>