inmates: arm: correct decoding of SGIs
If an IRQ is a SGI, bits 10-12 in the IAR of a GICv2 encode the
requesting CPU.

So far, Jailhouse doesn't set those bits (while it should, refer GICv2
spec!) so we didn't get hurt.  Nevertheless, this hits us when issuing
SGIs on bare-metal applications without jailhouse.  This patch properly
decodes SGI IRQ numbers by applying a mask to read_ack functions.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
[Jan: Removed unrelated hunk from header]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>