inmates: arm: fix erroneous interrupt stack address
The stack pointer must point to the top of the stack. Until now, it
erroneously pointed to the bottom of the stack.

Luckily this didn't lead to any problems so far, as the region before
the stack bottom was padded with zeroes.

Let's change the type of the irq_stack from u32 to u8. This allows us to
use simple stack top calculation with sizeof(), and without a nasty
ARRAY_SIZE workaround (which we would need for u32, and which we don't
for inmates).

Fixes: 9ee8e141893896 ("arm: basic inmates demos")
Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>