build: Avoid evaluating linux/compiler_types.h
Since 4.17, kbuild pulls in linux/compiler_types.h via -include in a way
we cannot avoid as long as we are using kbuild. This is not only against
our intention to avoid kernel headers for everything but the driver
module, it also breaks the build as we do not provide a search path for
the headers that compiler_types.h includes itself.

Fortunately, we can avoid the evaluation of that header simply by
defining its include guard.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>