build: Switch to lib.a for architecture parts of hypervisor
We already use this pattern for the inmate library. Moreover, our
current pattern of generating build-in.o no longer works with 4.17.
Switching to a library build is straightforward for arm and arm64, we
just need more explicit rules on x86 because of the need to build two
libraries in the same folder which is not directly supported by kbuild.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>