core: Fold map_root_memory_regions into init_late
It is safe to make this the very last step of init_late before
config_commit on all archs: it is only config_commit that depends on it,
and the region setup depends on the IOMMU being ready. And then we can
also fold this function into init_late directly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>