core: Introduce unit infrastructure
This introduces the concept of functional units which can register
themselves with the core.

Each unit will then receive a number of callbacks on events of common
interest: hypervisor init and shutdown, cell creation and destruction,
and counting of MMIO regions per cell. All callbacks are mandatory, so
unused ones need to be stubbed. For the more common cases of missing
shutdown or mmio_count_regions callbacks, macros are provided. The init
callbacks are run in linking order, cleanup callbacks in the reverse
order.

Registration is performed by adding a unit structure to the units
section. From there, for_each_unit* iterators can pick them up in
linking order (or reversed).

The goal of this abstraction is to remove explicit hooks for the
mentioned events for each unit. It also allows to add a unit by simply
linking it against the hypervisor. This is specifically useful for
existing target-specific variations (e.g. CAT for Intel or VExpress
extensions) or upcoming optional features.

The reason to call this "unit", instead of "module", is to avoid
confusion with abstractions in Linux or other projects that have broader
semantics. E.g., a unit is not intended to become a runtime loadable
entity.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>