core: Rename arch_shutdown to arch_prepare_shutdown
On x86, arch_shutdown now only contains preparation steps that all have
to be called prior to issuing the shutdown to units. Rename the function
and also the iommu service it calls on x86 to reflect that. ARM and
ARM64 do not use it, thus is unaffected by this.

No functional changes.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>