driver: Rework vpci overlay management for 4.17 - and before
As the API of 4.17 changes, replacing of_fdt_unflatten_tree +
of_overlay_apply with just of_overlay_fdt_apply, we have to adjust our
vpci overlay management. We now register the incomplete template, but in
a disabled state. Then we apply the required changes on the live tree,
including the activation of the node. This pattern nicely works for both
newer and older kernels.

Along this, we fix our lifecycle management of the the changeset as well
as its registered properties. The old code pushed in static properties,
but the OF core was implicitly expecting dynamically allocated objects.
Probably only due to resource leakage issues, we so far didn't trigger
any lethal attempt to kfree stack objects.

The new code has been validated extensively via kmemleak and kasan. Note
that, at the time of writing, the of_overlay_fdt_apply has a built-in
resource leakage that we cannot fix from this code. A solutions are
being discussed upstream.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>