configs: Adjust domains of virtual PCI controllers
We usually expect our virtual PCI controller to take domain 1 on
platforms that already have physical (but unvirtualizable) PCI host. If
we leave the domain at 0, we may accidentally unplug a real device from
the real host on shutdown. And if the virtual bridge should actually be
at domain 0, we will still unplug the device when we remove its host
controller's device tree overlay fragment.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>