arm64: Hide private per-cpu section during hypervisor operation
Turn off access to the private per-cpu section before the final
activation of the hypervisor. This prevents that one CPU can potentially
speculate about the private content of per-cpu sections of others.

We do not need to worry about the stack during the shutdown, it is
switched atomically with the MMU settings. But we still need to enable
access to the per-cpu section because arch_shutdown_self() will flush
it via its virtual address prior to turning off the MMU.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>