x86: Fix long mode detection
If the guest is currently running in long mode does not only depend on
EFER having it enabled. We also need to check if CS.L is set. If L is 0,
we are actually in compatibility mode.

Rework vcpu_vendor_get_cs() to vcpu_vendor_get_cs_attr() so that it
returns all needed information about the CS segment. This includes L,
the DPL for the current privilege level and already the D/B bit which
will be needed soon by the instruction parser.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>