x86: mmio: Fix address width calculation
As we are claiming to support the affected instructions, we should get
this completely right: When not in long mode, the combination of the
address-size prefix and the D/B bit from CS defines if we are using 16
or 32 bit addresses.

A similar pattern may eventually be useful to support the operand-size
prefix for with 16-bit accesses.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>