core: comm_region: Add magic and ABI revision to comm_region
Changes to the ABI of the communication region might break inmates
without noticing. Add the magic JHCOMM together with the ABI version as
a preamble of the structure.

The location signature, revision and cell_state are frozen from now on: their
location will be stable in future. Put those three fields at the beginning of
the structure, in this order.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>