core: reinitialise comm region on cell restart
The communication region is RW for both, the inmate and the hypervisor.
As a consequence, an inmate may overwrite the whole region. Currently,
the region is filled in cell_create.

Move initialisation to cell_start to provide a proper initial state when
starting the cell. Move architecture specific initialisation to
arch_cell_reset, which will be called by cell_start.

Leave the cell's shut down state in cell_create to maintain a valid
state after cell creation.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
[Jan: migrated cpu variable as well, massaged comment format in cell_start]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>