core: Fix paging_create when mapping single hugepages
When mapping hugepages, we try to clean up previous mappings of smaller
granularity for the target region first. Unfortunately, that could
destroy table entries in higher levels when the specific tree became
empty. This was easily reproducible by trying to map a single 2M region
at an address where there are no other mappings around.

The fix is to start the unmap with a subtree, rather than the full
paging tree. This way we will only clear entries at levels below the one
we are about to fill with a hugepage.

As it seems, the unmapping path is currently not used for practical
setups. The releasing impact could only by tested by artificially
defining two memory regions where the first one builds up a mapping with
3 levels and the second one overwrites this with only one level and a
hugepage. This case seems unlikely to occur with normal Jailhouse
reconfigurations. So we may consider dropping the cleanup path in the
future. It's fixed and kept for now to avoid surprises if the assumption
above no longer holds.

Reported-by: Giovani Gracioli <giovanig@gmail.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>