inmates: arm: move calculation of CONFIG_INMATE_BASE to inmate_common.h
We need to id-remap our own inmate. Therefore it needs to know its own
base address. Currently, the base address is determined by
CONFIG_INMATE_BASE in config.h.

So far, only linker scripts used this definition. Now we need it for
both: the linker script and in the code. Move it to inmate_common.h

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
[Jan: remove explicit config.h inclusion, indention tuning]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>