inmates: arm-common: add MMU support
Basically, the MMU on both ARM architectures work in the same way.

There are only slight differences in register naming and their content.
But the 'same' set of registers needs to be written. By abstracting
registers and their content to sysregs.h, we're now able to use the same
MMU driver code for both architectures.

Device drivers will ID-map the regions they are touching:
  - Base address of the inmate, length 0x10000 (results in 2MiB)
  - Communication Region, length 0x1000 (results in 2MiB)
  - UART, if existent, length 0x1000 (results in 2MiB)
  - GIC[CD|DR], length 0x1000 (results in 2MiB)

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>