configs: qemu-arm64: Set PCI domain
Upcoming QEMU 3.0 contains a change that injects linux,pci-domain into
the device tree it hands out to the guest. So we also need to set the
domain in the overlay for the virtual PCI host controller, otherwise
Linux refuses the fragment as invalid.

This change will make the config incompatible with older QEMU versions.
However, 3.0 will also contain some relevant fixes to the GIC emulation
so that lifting our version requirement is reasonable.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>