tools, pyjailhouse: Make installation of Python module and helpers optional
Depending on the availability of pip and, thus, also Python, install the
pyjailhouse module as well as the helpers. Warn otherwise, just like we
already do if python-mako is not available.

This reorganizes tools/Makefile so that Python-related rules go under
the ifeq and things that belong logically together are grouped
accordingly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>