core: Rework struct jailhouse_console
At the moment, the particular driver as well as its parameters are
encoded in the flags member of struct jailhouse_console.

Split it up in a type and a flags member. The type field selects the
particular driver, and the flags field annotates special options.

This not only makes configuration inside configuration files more clear,
it also saves fiddling around with bitmasks.

Instead of encoding the virtual console inside the flags field, encode
it in the global flags field. The virtual console is not a parameter for
a debug_console, like UART, it's rather a property of the whole cell.

This patch also renames JAILHOUSE_CON2_TYPE_ROOT_PAGE to
the JAILHOUSE_SYS_VIRTUAL_DEBUG_CONSOLE.

This patch
  - splits flags in type & flags in struct jailhouse_console
  - updates all system configurations
  - aligns the root cell config template

Patch for documentation will follow.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>