inmates: arm: Configure uarts via communication region
Everything is now in place, inmates are only missing the semantics of
struct jailhouse_console.

But this can be done easily by extending struct uart by a matching type
field that corresponds to jailhouse_console's type.

Passing con-type (et. al.) as inmate command line argument can still
override the configuration.

One con-type was removed on ARM: the "JAILHOUSE" con-type. On ARM, the
virtual console is now an additional secondary output path. It is
controlled via JAILHOUSE_CELL_VIRTUAL_CONSOLE_{PERMITTED/ACTIVE} in the
cell configuration and may be overridden by the boolean cmdline argument
"con-virtual".

The inmate will automatically write to that device if
JAILHOUSE_CELL_VIRTUAL_CONSOLE_ACTIVE is set in the configuration. The
permission bit is found in the communication region.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>