x86: Add EFI framebuffer debug console
All modern UEFI system come with the so-called Graphic Output Protocol
which provides a simple framebuffer device. This can also be used as
alternative to UART-based debug consoles. And it can replace the VGA
console which is no longer working on such systems (VGA is legacy).

As we are now filling a graphic buffer, we are in need of a console
font. The altc-8x16 we use here was taken from KBD version 2.0.4
(https://www.kernel.org/pub/linux/utils/kbd) and is licensed under
GPL-2.0-or-later. Leave a note about this also in the LICENSING file.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>