core: arm64: Fix gicv3_get_cpu_target
Shifting has precedence. But we need to shift _after_ we applied the
bitmask. Otherwise, the bitmask will applied after shifting the bits.
This is wrong.

This raises an error if we have more than eight CPUs: The eighth CPU
will be target 256, which will be masked out and result in 0.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>