x86: Wait on init_signaled to arrive
This particularly ensures that Linux guests do not consider CPUs dead
because of the potential high latency between submitting INIT and
getting the target CPU into wait_for_sipi state. On real hardware, the
Delivery Status bit which is polled by Linux synchronizes the sender.
Emulating this would be way more complex than making the INIT IPI
submission synchronous. Delaying this is fine because kicking off
secondary CPUs should never be a hot path.

In case of cross-posting between cores or when there is a management
request pending, we have to check for events and process them while
waiting for INIT to arrive.

The issue manifested in failing to online CPUs again under the root cell
Linux on an Intel NUC6CAY (Apollo Lake Atom).

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>