arm: Ensure double-word alignment of stack after entry
This issue was sleeping since day #1 of ARM support: The ARM calling
convention requires the stack to be double-word aligned on public
interfaces, e.g. when calling a subfunction. We failed to ensure this
both when calling entry() as well as arch_handle_exit().

This had no immediate negative impact, but was causing at least one
subtle effect: variable argument lists are expecting to find 64-bit
values double-word aligned on the stack, and that requirement was
violated. Therefore, printing long long variables was broken in the
hypervisor.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>