Splitting up helpers used in sysfs parsing
The information needed to determine if a machine can run jailhouse is located
entirely in sysfs.  To see if a machine can run jailhouse, the user must
currently (1) generate a configuration file using jailhouse-config-create (2)
run `make` (3) pass this file to jailhouse-hardware-check.

Prepare for enabling jailhouse-hardware-check to read this information directly
from sysfs by moving the relevant sysfs parsing functions in
jailhouse-config-create into a module in pyjailhouse.

Signed-off-by: Chris Goldsworthy <christopher.goldsworthy@outlook.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>