configs: Add support for MACCHIATObin board
Add configs for the Marvell Armada 8040-based MACCHIATObin board.

As we set the pci_domain to 1 and the upstream kernel does not
initialize linux,pci-domain, we need a patch to the DTB for the time
being.

MSI support should eventually be possible as well, but we still with
INTx for the virtual shmem devices for now.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>