x86: vmx: Intercept WBINVD instructions
WBINVD is a heavy instruction that write-backs and flushes all caches on
the source CPU socket, ignoring also CAT. Thus, this has the potential
of destroying worst-case latency assumptions of other cells. Linux, as a
primary guest of Jailhouse, only issues this instruction around CPU
offlining which we can safely ignore in the virtual environment.