configs: miriac-sbc-ls1046a: adjust memory flags for qbman portals
The qbman portal regions are split into two, cache-inhibited and
cache-enabled. (see DPAA_PORTAL_C[IE] in drivers/soc/fsl/qbman/)
When the underlying memory is not mapped correctly, like it was before
this commit, you will see TX errors and loose a significant amount of
packets on the DPAA NICs after jailhouse is enabled.

Signed-off-by: Henning Schild <henning.schild@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>