pyjailhouse: sysfs_parser: Fix wrong IVRS device ID for IOAPIC
If the type of an IVHD device entry is a Special Device (0x48),
then device ID is stored at a different location. The code already
parsed the value to device_id_b, but it was never used.

This set the ID of all IOAPICs to 0x0 in the system configuration.

Fixes: 5fe206927c05 ("tools: Implement ACPI IVRS table parses")

Signed-off-by: Andrej Utz <andrej.utz@st.oth-regensburg.de>
Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>