arm-common: gic-v3: Forward pending SGIs to virtual queue on enable
Analogously to gic-v2: When enabling Jailhouse, Linux may have some SGIs
pending that have to be taken off the physical interface and forwarded
to the virtual one. Otherwise, we would lose them or even confuse
Jailhouse because it expects SGIs to be injected virtually only (done on
interception of GICD_SGIR).

Based on original patch by Peng Fan.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>