Replicate and re-license printk-core
The inmate library currently uses printk-core.c from the hypervisor in
order to implement printf-like output. While the library has been
changed in 607251b44397 to dual-licensing, this part was ignored. So, if
an inmate uses printk, it can only choose GPL as license which is not
the intention of the re-licensing.

Change this situation by forking off the printk-core into the inmate
lib, now under dual BSD 2-clause / GPL. This is possible without hassle
as all contributions to hypervisor/printk-core.c were done by me, thus
are owned by Siemens. Forking avoids that people contributing to the
hypervisor core accidentally bring in GPL-only code to this part. It
also allows to evolve printk in the inmate library beyond what is needed
for the hypervisor without bloating the latter.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>