inmates: apic-demo: Plug race around message evaluation
If msg_to_cell was not JAILHOUSE_MSG_SHUTDOWN_REQUEST, we took the
default path which also sent back JAILHOUSE_MSG_UNKNOWN - and cleared
msg_to_cell before that. This created a race condition with the
hypervisor trying to sent the cell a message. Due to the hlt, the race
normally didn't matter so far. It will when moving to polling-based
idle.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>