inmates: apic-demo: Use polling idle to avoid latency spikes
It has been observed and also confirmed by Intel that hlt can cause
wake-up delay in the order of microseconds, although no particular power
management is requested this way. That's obviously because of
unpredictable decisions of the CPU how to react on that internally.

As the apic-demo is (also) about demonstrating the minimal possible
interrupt delay, switch to polling.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>