inmates: Make heap relocatable
The initial heap needs to be within the initially mapped memory in order
to allocate page tables from it for mapping more. On x86, growing the
heap beyond the first MB is blocked by the comm region that has to be
located in lower memory in order to avoid conflicts with PCI resources.
Therefore, it is needed to relocate the heap after mapping its
additional space. Inmates should be able to do that by setting heap_pos.
So, make this pointer global.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>